requirements.txt
streamlit
langchain
assemblyai
pytube
youtube-transcript-api
weaviate-client
openai
tiktoken